Inherit the legacy archiver, and make it run ifonly iOS or TvOS project
