+++ plugin/src/test/groovy/org/openbakery/XcodeBuildArchiveTaskOSXSpecification.groovy
@@ -12,7 +12,7 @@ class XcodeBuildArchiveTaskOSXSpecification extends Specification {
 
 	Project project
 
-	XcodeBuildArchiveTask xcodeBuildArchiveTask;
+	XcodeBuildLegacyArchiveTask xcodeBuildArchiveTask;
 
 	File projectDir
 	File buildOutputDirectory
